test(esl-utils): Extend `object#set` tests and split object utils tests between files
